Collect CatFacts out of json file
CatFacts (https://cat-fact.herokuapp.com/) is the public WebAPI which provides different facts about cats. Any user of the website can add a new fact, as well as to connect to this API and retrieve the facts wherever he/she wants it.
My code sends XML request to this public web API.
The Response Text is formatted in such a way, that there are no links with advertisements.
It makes convinient for the user to add the facts to the Excel Table / UserForm or wherever else. 
In order to make it work, the reference "Microsoft XML, v 6.0" must be added.